+++ examples/shop_app_example/lib/ui/theme.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:talker_flutter/talker_flutter.dart';
+import 'package:talker_shop_app_example/utils/good_log.dart';
 
 final lightTheme = ThemeData(
   primaryColor: primaryColor,
@@ -37,4 +38,6 @@ final Map<int, Color> _primaryColorCodes = {
   900: _primaryColor,
 };
 
-const talkerTheme = TalkerScreenTheme();
+final talkerTheme = TalkerScreenTheme(logColors: {
+  GoodLog.getKey: const Color(0xff4CAF50),
+});