+++ packages/talker/lib/src/extensions/extensions.dart
@@ -11,3 +11,12 @@ extension TalkerDataInterfaceListExt on List<TalkerData> {
     return sb.toString();
   }
 }
+
+extension TalkerIterableLogTypeModifier<TalkerLogType>
+    on Iterable<TalkerLogType> {
+  /// The method allows you to get the first element that satisfies the condition
+  /// or null if no element satisfies the condition.
+  TalkerLogType? firstWhereOrNull(bool Function(TalkerLogType element) test) =>
+      cast<TalkerLogType?>()
+          .firstWhere((v) => v != null && test(v), orElse: () => null);
+}