+++ packages/talker_flutter/example/lib/extended_example/extended_example.dart
@@ -84,7 +84,7 @@ class _ExtendedExampleState extends State<ExtendedExample> {
   }
 
   void _customLog() {
-    _talker.logTyped(CustomLog('Custom log message'));
+    _talker.logCustom(CustomLog('Custom log message'));
   }
 
   void _criticalLog() {