+++ packages/talker_flutter/lib/src/ui/theme/talker_screen_theme.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:talker_flutter/src/ui/theme/default_theme.dart';
 import 'package:talker_flutter/talker_flutter.dart';
 
-typedef LogColors = Map<TalkerLogType, Color>;
+typedef LogColors = Map<String, Color>;
 
 /// Configuring the UI of [TalkerScreen]
 class TalkerScreenTheme {
@@ -22,11 +22,11 @@ class TalkerScreenTheme {
   /// Color of [Talker] data cards
   final Color cardColor;
 
-  final Map<TalkerLogType, Color>? _customColors;
+  final Map<String, Color>? _customColors;
 
-  Map<TalkerLogType, Color> get logColors {
+  Map<String, Color> get logColors {
     if (_customColors != null) {
-      final customMap = Map<TalkerLogType, Color>.from(_defaultColors);
+      final customMap = Map<String, Color>.from(_defaultColors);
       customMap.addAll(_customColors!);
       return customMap;
     }
@@ -45,31 +45,31 @@ class TalkerScreenTheme {
 
 extension MapTalkerFlutterColorsExt on LogColors {
   Color getByType(TalkerLogType type) {
-    return this[type] ?? Colors.grey;
+    return this[type.key] ?? Colors.grey;
   }
 }
 
-const _defaultColors = {
+final _defaultColors = {
   /// Base logs section
-  TalkerLogType.error: Color.fromARGB(255, 239, 83, 80),
-  TalkerLogType.critical: Color.fromARGB(255, 198, 40, 40),
-  TalkerLogType.info: Color.fromARGB(255, 66, 165, 245),
-  TalkerLogType.debug: Color.fromARGB(255, 158, 158, 158),
-  TalkerLogType.verbose: Color.fromARGB(255, 189, 189, 189),
-  TalkerLogType.warning: Color.fromARGB(255, 239, 108, 0),
-  TalkerLogType.exception: Color.fromARGB(255, 239, 83, 80),
+  TalkerLogType.error.key: const Color.fromARGB(255, 239, 83, 80),
+  TalkerLogType.critical.key: const Color.fromARGB(255, 198, 40, 40),
+  TalkerLogType.info.key: const Color.fromARGB(255, 66, 165, 245),
+  TalkerLogType.debug.key: const Color.fromARGB(255, 158, 158, 158),
+  TalkerLogType.verbose.key: const Color.fromARGB(255, 189, 189, 189),
+  TalkerLogType.warning.key: const Color.fromARGB(255, 239, 108, 0),
+  TalkerLogType.exception.key: const Color.fromARGB(255, 239, 83, 80),
 
   /// Http section
-  TalkerLogType.httpError: Color.fromARGB(255, 239, 83, 80),
-  TalkerLogType.httpRequest: Color(0xFFF602C1),
-  TalkerLogType.httpResponse: Color(0xFF26FF3C),
+  TalkerLogType.httpError.key: const Color.fromARGB(255, 239, 83, 80),
+  TalkerLogType.httpRequest.key: const Color(0xFFF602C1),
+  TalkerLogType.httpResponse.key: const Color(0xFF26FF3C),
 
   /// Bloc section
-  TalkerLogType.blocEvent: Color(0xFF63FAFE),
-  TalkerLogType.blocTransition: Color(0xFF56FEA8),
-  TalkerLogType.blocClose: Color(0xFFFF005F),
-  TalkerLogType.blocCreate: Color.fromARGB(255, 120, 230, 129),
+  TalkerLogType.blocEvent.key: const Color(0xFF63FAFE),
+  TalkerLogType.blocTransition.key: const Color(0xFF56FEA8),
+  TalkerLogType.blocClose.key: const Color(0xFFFF005F),
+  TalkerLogType.blocCreate.key: const Color.fromARGB(255, 120, 230, 129),
 
   /// Flutter section
-  TalkerLogType.route: Color(0xFFAF5FFF),
+  TalkerLogType.route.key: const Color(0xFFAF5FFF),
 };