+++ packages/talker/CHANGELOG.md
@@ -1,3 +1,42 @@
+# 4.5.1
+- [talker_flutter] Fix mapping LogColors to Flutter Color (Logs screen)
+
+# 4.5.0
+More Flexible Interaction with Custom Logs  
+- **BREAKING** [talker] Change the type of `colors` parameter in the `TalkerSettings` class from `Map<TalkerLogType, AnsiPen>?` to `Map<String, AnsiPen>?`. Custom colors must now use a string key for log types.  
+- **BREAKING** [talker] Change the type of `titles` parameter in the `TalkerSettings` class from `Map<TalkerLogType, String>?` to `Map<String, String>?`. Custom titles must now use a string key for log types.  
+- **BREAKING** [talker_flutter] Change the type of `colors` parameter in the **TalkerScreenTheme** class from `Map<TalkerLogType, Colors>?` to `Map<String, Colors>?`. Custom colors must now use a string key for log types.  
+- [talker] Add new tests and updated existing ones.  
+- [talker] Update the documentation for log customization.
+
+Thanks to [yelmuratoff](https://github.com/yelmuratoff) 
+
+# 4.4.7
+- [talker_dio_logger] Make encoder constant private
+
+# 4.4.6
+- [talker] Reanme **logTyped** method to **logCustom**
+- [talker] Add **Deprecated** annotation for logTyped method
+
+# 4.4.5
+- [talker_logger] Add enable option field to control log output (on talker_logger level)
+
+Thanks to [weitsai](https://github.com/weitsai)
+
+# 4.4.4
+- [talker_dio_logger] Add enable option field to control log output (on talker_dio_logger level)
+
+Thanks to [weitsai](https://github.com/weitsai)
+
+# 4.4.3
+- [talker_http_logger] Bump http_interceptor package version to **2.0.0**
+
+# 4.4.2
+- [talker] Fix provide pen color field in log() method
+- [talker] Android dependecies actualization and bug fixing in example app code
+
+Thanks to [HE-LU](https://github.com/HE-LU) and [venkat9507](https://github.com/venkat9507)
+
 # 4.4.1
 - [talker_flutter] Bump share_plus version to 10.0.1
 - [talker_flutter] Bump path_provider version to 2.1.4